Rename and Refactor CalciteLogicSchemaFactory
